pci: Move errors into their own module
We'll need to use the Error and Result types outside of just probing the
device-tree.
